refactor(console): modularize App.tsx and fix navigation dropdown blur logic
